syscall: extend chroot check in unit tests
Extend the current check (which checks if the inode of / is 2)
by one comparing `/proc/{1,self}/mountinfo`.

This check may not always work, which is why the original inode-based
check is left intact. This essentially makes the result of `isChroot`
identical to running the `ischroot` command on systems that provide it.